requirements.txt
pycryptodome
black
sh
git+https://github.com/blockstack/secret-sharing

